==> work_trial/step02_dimensionality_reduction.ipynb <==
# Introduction

This notebook reports the Pearson correlation between each variable and the target outcome. While these correlations cannot tell us exactly which variables will offer the best performance in the context of a holistic productionized model, they're a good starting point for dimensionality reduction. 

import numpy as np
import pandas as pd
import scipy.stats

from utils import bad_rate_by_category

pd.set_option('display.max_rows', 200)
pd.set_option('display.max_columns', 50)

## Data

# Import data and data dictionary
data = pd.read_pickle('output_data/01_data.pkl')

data_dict = pd.read_pickle('output_data/01_data_dict.pkl')

## Potential Features

# Designate eda categories as containing features
prelim_feature_dict = ['credit_score', 'personal_finance', 'other_info']
data_dict['potential_feature'] = data_dict.apply(lambda x: x.eda_category in prelim_feature_dict, axis=1)

## Pearson Correlations

# Function to evaluate pearson correlation
def correlate_pearson(var1, var2, data):
    # Drop rows without either input or output variable
    data_notna = data.dropna(axis=0, subset=[var1, var2])
    
    # Run Pearson correlation
    pearson, pearson_p = scipy.stats.pearsonr(data_notna[var1], data_notna[var2])
    
    return pearson, pearson_p

# Designate list of potential model features
potential_features = data_dict.loc[data_dict['potential_feature']==True, 'variable'].values

# For each potential feature, calculate Pearson's correlation 
# and its p-value and store them in the data dictionary
for var in potential_features:
    pearson, pearson_p = correlate_pearson(var, 'bad', data)
    data_dict.loc[data_dict['variable']==var,'pearson'] = pearson
    data_dict.loc[data_dict['variable']==var,'pearson_p'] = pearson_p

data_dict.sort_values('pearson_p').reset_index(drop=True)

## Export Data

data.to_pickle('output_data/02_data.pkl')
data_dict.to_pickle('output_data/02_data_dict.pkl')
