==> work_trial/step03_logistic_regression_incomplete.ipynb <==
# Did Not Complete!

Using one model for a modeling project is never a good idea, especially when it's a complex model like XGBoost. You want to have straightforward and easily interpretable models established as a baseline against which to compare your fancy models which have (hopefully) better performance. 

With limited time to complete this work trial project, I skipped the logistic regression and went straight to XGBoost. But it's an important step in my process and I would always finish this before deploying or making recommendations to a client. Therefore, I've included this placeholder notebook to indicate the full plan.
