==> work_trial/step05_performance_evaluation.ipynb <==
# Introduction

Model selection deserves more critical thinking than choosing the best model on a single statistical metric. The model selection process should include analyzing the impact on approvals and interest rates and seeing which protected demographic populations benefit by which models.

This notebook begins the model performance evaluation, using data generated by the sklearn hyperparameter grid search. It compares the performance of the model on key metrics for each possible hyperparameter selection to aid in model selection.

import os
import numpy as np
import pandas as pd
import scipy.stats
import seaborn as sns
import matplotlib.pyplot as plt

import multiprocessing

import matplotlib.pyplot as plt

pd.set_option('display.max_rows', 500)

## Data

Let's load up the results of the last notebook. Because we're building an alternative credit score, we'll only consider models that do not use FICO as an input. 

# Load performance results from previous notebook 

# Store everything in one performance dataframe
performance = pd.DataFrame()

# Iterate through feature sets and search numbers
feature_sets = ['no_credit']
search_numbers = ['0001', '0002', '0003']

# For each search
for i in search_numbers:
    # For each feature set
    for name in feature_sets:
        # Store data in a dictionary
        df = pd.read_pickle(f'output_data/04_{name}_performance_{i}.pkl')
        
        # Track the features it used and the grid search it came from
        df['features'] = name
        df['grid_search'] = i
        
        # Append to the overall performance DF
        performance = performance.append(df)

#### Let's rename some of these columns. 

performance.rename({'mean_fit_time': 'training_seconds'},
                   inplace = True)

#### Build some new columns. 

performance = performance.reset_index()
performance = performance.rename({'index': 'model_id'}, axis=1)

performance['framework'] = 'xgboost'

#### Take a look. 

performance.head()

#### Build a new Performance Dictionary

performance_dict = pd.DataFrame(performance.dtypes)
performance_dict = performance_dict.reset_index()
performance_dict.rename({0: 'var_dtype',
                        'index': 'var_name'}, 
                        axis=1, 
                        inplace = True)

performance_dict.head()

performance_dict.head()

#### Categorize Variables with NLP Regex Tagging

category_tags = {
    '_time': 'runtime',
    'param_': 'model_parameters',
    '_test_': 'score' 
    # Left is Data (including aggregated data) 
    # Right is eval_metric
}

for k, v in category_tags.items():
    performance_dict.loc[performance_dict['var_name'].str.contains(k),
                        'performance_category'] = v

# Evaluate Hyperparameter Performance

# Designate Hyperparameters
hypers = ['param_subsample', 
          'param_max_depth', 
          'param_min_child_weight',
          'param_reg_lambda',
          'param_gamma', 
          'param_eta',
          'features']

def hyper_results(performance: pd.DataFrame, hyper: str):
    """
    Summarizes the effects of each hyperparameter on validation set scoring metrics. 
    
    """
    groupby_dict = {
   
    }
    gb = performance.groupby(hyper).agg({'mean_test_accuracy': 'mean',
                                    'mean_test_neg_brier_score': 'mean',
                                    'mean_test_roc_auc': 'mean',
                                    'mean_test_precision': 'mean',
                                    'mean_test_recall': 'mean'})
    
    
    return gb

#### Maximum Depth of a Tree
Increasing the max depth of the tree has little effect on most measures of accuracy. Deeper trees have higher precision and lower recall scores. 

hyper_results(performance, 'param_max_depth')

#### Minimum Sample Size for a New Node
You can strongly improve performance by increasing the minimum sample size for a new tree leaf. The best models we have use the highest possible min_child_weight value, so we should run more tests to find out if higher weights would be even more performant. 

hyper_results(performance, 'param_min_child_weight')

#### Minimum Performance Gain for a New Node
Gamma is good! Raise to 10, 20, and 50 to compare. 

hyper_results(performance, 'param_gamma')

#### Lambda (L2) Regularization Parameter
To prevent overfitting, you can use an L1 regularization parameter. But in this test, it has no visible effect on performance. 

hyper_results(performance, 'param_reg_lambda')

#### Learning Rate (Eta)

Prevents overfitting by minimizing the impact of new trees on the overall prediction. Over time, adjustments become fine-grained. 

'eta': [0.05, 0.005, 0.001, 0.00001], 

hyper_results(performance, 'param_eta')

#### Subsample %
You can also fiddle with dropout layers / subsampling within the model, but no big effects here. 

hyper_results(performance, 'param_subsample')

# Selecting One Model

select_cols = [
    'model_id',
    'param_booster',
    'param_max_depth',
    'param_min_child_weight',
    'param_reg_lambda',
    'param_gamma',
    'param_eta',
    'param_subsample',
    'mean_test_accuracy',
    'mean_test_neg_brier_score',
    'mean_test_roc_auc',
    'mean_test_precision',
    'mean_test_recall',
    'features'
]

top_100 = performance[select_cols].sort_values('mean_test_neg_brier_score', ascending=False).head(100)

for col in hypers:
    print(top_100[col].value_counts())

# Choose by hand the most effective hyperparameters for the model
best_hypers = {
    'param_max_depth': 4,
    'param_min_child_weight': 16,
    'param_reg_lambda': 0,
    'param_gamma': 10,
    'param_eta': 0.10,
    'param_subsample': 0.7,
}
